% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2. backwards_selection.R
\name{backward_selection}
\alias{backward_selection}
\title{Backwards Variable Selection for Random Forests.}
\usage{
backward_selection(X, y, ntrees = 1000, plot = F, output = T)
}
\arguments{
\item{X}{A data frame or matrix with the for each covariate a column .}

\item{y}{A vector of response.}

\item{ntrees}{The number of trees estimated in one step}

\item{plot}{boolean indicating if a plot should be made of the entire procedure}

\item{output}{boolean whether to print progress of the selection}
}
\value{
vars the chosen variables of the optimal model, the prediction error of the chosen model, selection_results of each step
}
\description{
Backwards Variable Selection for Random Forests.
}
