% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2. backwards_selection.R
\name{backwards_step}
\alias{backwards_step}
\title{One Backwards Variable Selection Step.}
\usage{
backwards_step(data, ntrees)
}
\arguments{
\item{data}{a dataframe with y and X}

\item{ntrees}{The number of trees estimated in one step}
}
\value{
var_drop the variable with the lowest importance score, pred_error the prediction error of the model, pred_error_sd the prediction error standard deviation
}
\description{
One Backwards Variable Selection Step.
}
