% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Hyperspectral data acquired over a forest area}
\format{
A data frame with 3230 rows and 66 variables
}
\usage{
data(dat)
}
\description{
A dataset containing 3230 samples with 65 hyperspectral bands and 8 classes.
}
\details{
\itemize{
  \item B1...B65 Hyperspectral bands.
  \item SP. Classes.
}
}
\keyword{datasets}
