% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMdist.R
\name{JMdist}
\alias{JMdist}
\title{Jeffries-Matusita distance among classes}
\usage{
JMdist(g, X)
}
\arguments{
\item{g}{A column vector of the lables. length(g) is equal to nrow(X).}

\item{X}{A dataframe of the features. ncol(X) is equal to the total number of features, and nrow(X) is equal to the number of avaialble training samples. nrow(X) is equal to length(g)}
}
\value{
A list containing a matrix of the class combinations and a vector of the JM distances of all the class combinations.
}
\description{
Jeffries-Matusita distance.
}
\author{
Michele Dalponte and Hans Ole Oerka
}
\references{
Dalponte, M., Oerka, H.O., Gobakken, T., Gianelle, D. & Naesset, E. (2013). Tree Species Classification in Boreal Forests With Hyperspectral Data. IEEE Transactions on Geoscience and Remote Sensing, 51, 2632-2645.
}

