% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. forward_selection.R
\name{quant_error}
\alias{quant_error}
\title{calculation of the quantile error for random forest fit}
\usage{
quant_error(fit, y, taus)
}
\arguments{
\item{fit}{A fitted gradient random forest object}

\item{y}{A vector of observations used for fit}

\item{taus}{A vector of probability levels used for calculation of the quantile error}
}
\value{
The sum of the quantile error for each of the probability levels from taus calculated from the Out-Of-Bag samples
}
\description{
calculation of the quantile error for random forest fit
}
