% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. forward_selection.R
\name{select_sample_fraction}
\alias{select_sample_fraction}
\title{Selection of the optimal sub sample size for the variable selection}
\usage{
select_sample_fraction(X, y, sf_seq = seq(0.05, 0.25, 0.05),
  ntrees = 50)
}
\arguments{
\item{X}{The data matrix of covariates}

\item{y}{A vector of observations used for fit}

\item{sf_seq}{a sequence of subsample fractions to try}

\item{ntrees}{Number of trees for the variable selection procedure}
}
\value{
The optimal subsample fraction
}
\description{
Selection of the optimal sub sample size for the variable selection
}
