% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. forward_selection.R
\name{varSel}
\alias{varSel}
\title{Forward Variable Selection for Random Forests.}
\usage{
varSel(X, y, nsteps = NULL, model_vars = c(), taus = seq(0.1, 0.9,
  0.1), alpha = 0.95, ntrees = 500, sf = 0.5, mtry = 10,
  min_node_size = 1, select_sf = F, output = T, num.threads = 2)
}
\arguments{
\item{X}{A data frame or matrix with the for each covariate a column .}

\item{y}{A vector of response.}

\item{nsteps}{The number of selection steps that need to be done if NULL new variables are added until no more variables are available}

\item{taus}{The sequence of probability levels, the corresponding quantiles are used to evaluate the performance of the model}

\item{alpha}{The confidence level for the test default set to 95 percent}

\item{ntrees}{The number of trees estimated in one step}

\item{sf}{The subsample fraction that is used in the estimation}

\item{output}{Indicating wheter every forward step the function prints which variable is selected}

\item{model.vars}{The variables that the model will start with}
}
\value{
The selected variables in model_vars and the selection results in selection.results.
}
\description{
Forward Variable Selection for Random Forests.
}
