% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. forward_selection.R
\name{varSel_step}
\alias{varSel_step}
\title{One foreward step for variable selection}
\usage{
varSel_step(X, y, model_vars, taus, ntrees, sf, mtry, min_node_size,
  num.threads)
}
\arguments{
\item{X}{A data frame or matrix with the for each covariate a column}

\item{y}{A vector of response.}

\item{model_vars}{The variables that the model will start with}

\item{taus}{The sequence of probability levels, the corresponding quantiles are used to evaluate the performance of the model}

\item{ntrees}{The number of trees estimated in one step}

\item{sf}{The subsample fraction that is used in the estimation}

\item{mtry}{the number of variables to consider at each split}

\item{min_node_size}{the minimum number of vairables in each leaf node}
}
\value{
Data frame with a row with the calculated quantile loss and a row indicating which variable. The data frame is sorted on the quantileLoss
}
\description{
One foreward step for variable selection
}
